Do not start processing before all threads have been created
Process-creation create a lot of kernel-work and will induce jitter in
the scheduling. We are not measuring this kind of jitter and should
avoid this at all cost (false positives, if you like).

This will cause semtest's marcos to wait for the master thread to signal
readiness before proceeding, hopefully removing this jitter from the
output.

Signed-off-by: Henrik Austad <haustad@cisco.com>